+++ kmpotest/kmpotest/src/jsMain/kotlin/at/asitplus/test/JUnitXMLReporter.kt
@@ -1,3 +1,12 @@
 package at.asitplus.test
 
-internal actual val target: String ="js"
+internal actual val target: String ="js${detectEnvironment()}"
+
+fun detectEnvironment(): String {
+    return when {
+        js("typeof window !== 'undefined'") as Boolean -> "Browser"
+        js("typeof global !== 'undefined'") as Boolean -> "Node"
+        js("typeof self !== 'undefined'") as Boolean -> "WebWorker"
+        else -> "unknown"
+    }
+}